Some volume related sound funcs

+++ include/functions.us.h
@@ -58,7 +58,7 @@ void func_800F7F50_2F5A30(void);
 void func_8000941C(unkObjectStruct*, f32, f32, f32);
 void func_800090A4(unkObjectStruct*);
 void func_80009448(void);
-void func_8000C250(s32);
+void func_8000C250(s16);
 void func_8004A510(void);
 void func_8004CC8C(s16, s16);
 void func_8004CDA0(s16);
